+++ src/AssetManager/Audio.cpp
@@ -25,7 +25,7 @@ namespace AssetManager
 		int sampleRate;
 		short* audioBuffer;
 		int dataLength;
-		dataLength = stb_vorbis_decode_memory(static_cast<const uint8*>(data), size, &channels, &sampleRate, &audioBuffer);
+		dataLength = stb_vorbis_decode_memory(static_cast<const uint8*>(data), static_cast<int>(size), &channels, &sampleRate, &audioBuffer);
 
 		alGenBuffers(1, &buffer);
 		alBufferData(buffer, AL_FORMAT_MONO16, audioBuffer, dataLength * sizeof(ALshort), sampleRate);
